Fix a bunch of Eclipse warnings

+++ imglib2/core/src/test/java/net/imglib2/BinningTest.java
@@ -436,7 +436,7 @@ public class BinningTest {
 
         // fill up the histogram with uniform values
         for (double value = min; value <= max; value += inc) {
-            int bin = (int) Binning.valueToBin(bins, min, max, value);
+            int bin = Binning.valueToBin(bins, min, max, value);
             ++histogram[bin];
         }
